Move sensors.py to `sensors/base_sensor.py`
