+++ ami/environments/sensors/__init__.py
@@ -1,0 +1,1 @@
+from . import base_sensor
